dom/security/test/csp/test_upgrade_insecure_loopback.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Bug 1447784 - Implement CSP upgrade-insecure-requests directive</title>
  <!-- Including SimpleTest.js so we can use waitForExplicitFinish !-->
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<iframe style="width:100%;" id="testframe"></iframe>

<script class="testbody" type="text/javascript">

/* Description of the test:
 * We load a page that performs a CORS XHR to 127.0.0.1 which shouldn't be upgraded to https:
 *
 * Test 1:
 *   Main page:   https://127.0.0.1:8080
 *   XHR request: http://127.0.0.1:8080
 *   No redirect to https://
 *   Description: Upgrade insecure should *NOT* upgrade from http to https.
 */

const CSP_POLICY = "upgrade-insecure-requests; script-src 'unsafe-inline'";
let tests = 1;

function loadTest() {
  var src = "https://example.com/tests/dom/security/test/csp/file_testserver.sjs?file=";
  // append the file that should be served
  src += escape("tests/dom/security/test/csp/file_upgrade_insecure_loopback.html")
  // append the CSP that should be used to serve the file
  src += "&csp=" + escape(CSP_POLICY);
  document.getElementById("testframe").src = src;
}

function checkResult(result) {
  if (result === "request-not-https") {
    ok(true, "Didn't upgrade 127.0.0.1:8080 to https://");
  }
  if (--tests > 0) {
    return;
  }
  window.removeEventListener("message", receiveMessage);
  SimpleTest.finish();
}

// a postMessage handler that is used to bubble up results from
// within the iframe.
window.addEventListener("message", receiveMessage);
function receiveMessage(event) {
  checkResult(event.data);
}

SimpleTest.waitForExplicitFinish();

// By default, proxies don't apply to 127.0.0.1.
// We need them to for this test (at least on android), though:
SpecialPowers.pushPrefEnv({set: [
  ["network.proxy.allow_hijacking_localhost", true]
]}).then(loadTest);

</script>
</body>
</html>
